Extract peer-link session machinery into session.py
Sixth and final PR in the peer_link split arc (follows #769 +
#770 + #773 + #781 + #783). Moves the long-lived
post-handshake session machinery byte-for-byte from
peer_link/__init__.py into peer_link/session.py:

* PeerLinkSession dataclass + send_app_frame / terminate
* _run_peer_link_session driver
* _receive_loop frame dispatcher
* parse_app_frame validation seam (shared with offloader-side)
* run_peer_link_heartbeat (drives either end)
* _monotonic clock indirection (test hook)
* Constants: HEARTBEAT_INTERVAL_SECONDS,
  HEARTBEAT_MISS_THRESHOLD, HEARTBEAT_DEAD_AFTER_SECONDS,
  APP_FRAME_MAX_BYTES

Lazy imports cleaned up:
* handshake.py's _run_peer_link_session lazy import becomes
  a module-level from .session import.
* channel.py's lazy parse_app_frame import redirects to
  .session (channel.py is imported BY session.py, so this
  one stays lazy to avoid circular load).

__init__.py drops the heartbeat + frame-cap constant blocks
(both moved to session.py) and the section-divider comment.
Its remaining content is just the public WS handler factory
(make_peer_link_handler), PEER_LINK_PATH, and re-exports that
keep external imports stable.

Test redirects (the HIGH-risk part of this PR):
* _peer_link_module monkey-patches against HEARTBEAT_*,
  _monotonic, run_peer_link_heartbeat redirect to a new
  _peer_link_session_module reference.
* Direct calls of _peer_link_module.run_peer_link_heartbeat
  redirect to _peer_link_session_module.
* PeerLinkNoiseSession.responder patch redirects from
  peer_link.PeerLinkNoiseSession to
  peer_link.handshake.PeerLinkNoiseSession.
* _receive_loop test import redirects to
  peer_link.session._receive_loop.

__init__.py: 519 -> 144 lines (-375). session.py: 398 lines.

Full peer_link arc: 1071 lines in one file -> six cohesive
submodules with __init__.py at 144 lines (just the public WS
factory + re-exports).

All 3391 non-e2e tests pass; ruff + mypy clean.
